requirements.txt
# Walmart Outlook Importer Dependencies
# Python 3.13.9

# Browser automation
playwright>=1.40.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Environment variables
python-dotenv>=1.0.0

# Windows COM for Outlook
pywin32>=306

# Optional: Retry utilities
tenacity>=8.2.0
